# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    Prediction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        marketId:
          type: string
        status:
          type: string
          enum: &ref_0
            - pending
            - confirmed
            - won
            - lost
            - claimed
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - marketId
        - status
        - createdAt
...
    PredictionRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        marketId:
          type: string
        question:
          type: string
        outcome:
          type: string
        amount:
          type: string
        txHash:
          type: string
        status:
          type: string
          enum: *ref_0
        result:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        resolutionTime:
          type: string
          format: date-time
      required:
        - id
        - marketId
        - question
        - outcome
        - amount
        - txHash
        - status
        - result
        - createdAt
        - resolutionTime
    PredictionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PredictionRow'
        nextCursor:
          type: string
          nullable: true
      required:
        - data
        - nextCursor
...
paths:
...
  /api/users/{address}/predictions:
    get:
      operationId: getUserPredictions
      tags:
        - Users
      summary: List predictions for a Stellar address
      parameters:
        - schema:
            type: string
          required: true
          name: address
          in: path
        - schema:
            type: string
            enum: *ref_0
          required: false
          name: status
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated predictions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prediction'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
              examples:
                samplePage:
                  value:
                    data:
                      - id: 11111111-1111-1111-1111-111111111111
                        marketId: market-abc-123
                        status: confirmed
                        createdAt: '2026-06-27T12:00:00.000Z'
                    nextCursor: djF8MjR8...
        '400':
          description: Invalid address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                invalidAddress:
                  value:
                    error:
                      code: invalid_address
                      requestId: req_abc123
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                notFound:
                  value:
                    error:
                      code: not_found
                      requestId: req_abc123
...
  /api/predictions:
    get:
      operationId: listPredictions
      tags:
        - Predictions
      summary: List the authenticated user’s predictions
      description: >-
        Returns a cursor-paginated list of predictions placed by the caller. Sort order is `createdAt DESC, id DESC`.
        Pass the returned `nextCursor` as `?cursor=` to fetch the next page. `nextCursor` is `null` when no further
        pages exist.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 128
          required: false
          name: marketId
          in: query
        - schema:
            type: string
            enum: *ref_0
          required: false
          name: status
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 64
          required: false
          name: outcome
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated list of predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionsListResponse'
              examples:
                authenticatedPredictionsPage:
                  value:
                    data:
                      - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                        marketId: market_123
                        question: Will Bitcoin hit 100k in 2026?
                        outcome: 'yes'
                        amount: '50.0000000'
                        txHash: 8c253240be423ef8109d94101e40a02bc8f297b819f0ff4f4c20b8e906059e66
                        status: won
                        result: 'yes'
                        createdAt: '2026-05-01T12:00:00.000Z'
                        resolutionTime: '2026-06-01T12:00:00.000Z'
                    nextCursor: cursor_abc123
        '400':
          description: Validation error — invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
              examples:
                invalidLimit:
                  value:
                    error:
                      code: VALIDATION_ERROR
                      details: Limit must be between 1 and 100
        '401':
          description: Unauthorized — missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                unauthorized:
                  value:
                    error:
                      code: UNAUTHORIZED
                      requestId: req_xyz789
...